Initial support for HAL_PIXEL_FORMAT_TI_NV12
Both solid fill and file sources are supported. Files should have
the native stride for best performance, but other strides are
supported as well.

For video simulation, setting a surface to async mode is supported
with the async flag in the test spec. NV12 is specified by setting
buffer_format to either HAL_PIXEL_FORMAT_TI_NV12 or 256 (0x100)

Change-Id: Icefc7ee7f8e75133486cb3933f3279072db1a7b6
Reviewed-on: http://gerrit.nsc.com/177
Reviewed-by: Dima Svetlov <a0274124@ti.com>
Tested-by: Dima Svetlov <a0274124@ti.com>