More visually appealing surface resize and movement
Instead of restoring position and size to initial value
when limits are reached, grow and shrink / bounce between
min and max value. For size this gives a pulsating effect
and for position a bouncing effect.

This has no real impact on actual test cases but makes
them more visually appealing.

Change-Id: I96fd6498b0e3839cb5cb70f977d56d464f5ba351
Reviewed-on: http://gerrit.nsc.com/249
Reviewed-by: Jonas Larsson <jonas.larsson@ti.com>
Tested-by: Jonas Larsson <jonas.larsson@ti.com>
Reviewed-by: Dima Svetlov <a0274124@ti.com>