Duty cycles for size & position, default values
Since a resize operation always triggers a content
update we need a duty cycle for it to avoid
constantly repainting the surface each iteration when
resizing is enabled.

This also updates the format for duty cycle to:
    property_name on off
This saves one line per duty cycle in the case spec.
Old format is still parsed as well to keep
compatibility. Already released examples still work.

Test specs are now created with initial values for all
properties to avoid complete weirdness if user forgets
to set one.

Change-Id: Ibf743faa9910195539b7ff3ea55a1440d903699b
Reviewed-on: http://gerrit.nsc.com/258
Reviewed-by: Dima Svetlov <a0274124@ti.com>
Tested-by: Dima Svetlov <a0274124@ti.com>